+++ src/test/java/kitchenpos/api/ordertable/TableChangeNumberOfGuestsApiTest.java
@@ -1,7 +1,7 @@
-package kitchenpos.api.table;
+package kitchenpos.api.ordertable;
 
 import kitchenpos.api.config.ApiTestConfig;
-import kitchenpos.ordertable.application.dto.request.OrderTableChangeEmptyRequest;
+import kitchenpos.ordertable.application.dto.request.OrderTableChangeGuestNumberRequest;
 import kitchenpos.ordertable.application.dto.response.OrderTableResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -14,25 +14,25 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-class TableChangeEmptyApiTest extends ApiTestConfig {
+class TableChangeNumberOfGuestsApiTest extends ApiTestConfig {
 
-    @DisplayName("주문 테이블 주문 가능 상태로 변경 API 테스트")
+    @DisplayName("주문 테이블 주문자 수 변경 API 테스트")
     @Test
-    void changeEmptyTable() throws Exception {
+    void changeNumberOfGuestsTable() throws Exception {
         // given
-        final Long orderTableId = 1L;
-        final OrderTableChangeEmptyRequest request = new OrderTableChangeEmptyRequest(false);
+        final OrderTableChangeGuestNumberRequest request = new OrderTableChangeGuestNumberRequest(4);
 
         // when
-        final OrderTableResponse response = new OrderTableResponse(orderTableId, 1, request.getEmpty(), null);
-        when(tableService.changeEmpty(eq(orderTableId), eq(request))).thenReturn(response);
+        final Long orderTableId = 1L;
+        final OrderTableResponse response = new OrderTableResponse(orderTableId, request.getNumberOfGuests(), false, null);
+        when(tableService.changeNumberOfGuests(eq(orderTableId), eq(request))).thenReturn(response);
 
         // then
-        mockMvc.perform(put("/api/tables/{id}/empty", orderTableId)
+        mockMvc.perform(put("/api/tables/{id}/number-of-guests", orderTableId)
                         .contentType(APPLICATION_JSON_VALUE)
                         .content(objectMapper.writeValueAsString(request)))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.id", is(response.getId().intValue())))
-                .andExpect(jsonPath("$.empty", is(response.isEmpty())));
+                .andExpect(jsonPath("$.numberOfGuests", is(response.getNumberOfGuests())));
     }
 }
