+++ src/test/java/kitchenpos/api/tablegroup/TableGroupCreateApiTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.api.table;
+package kitchenpos.api.tablegroup;
 
 import kitchenpos.api.config.ApiTestConfig;
 import kitchenpos.tablegroup.application.dto.request.TableGroupCreateRequest;
